+++ src/servidor/modelo/movimiento/Posicion.cpp
@@ -1,6 +1,6 @@
 #include "includes/servidor/modelo/movimiento/Posicion.h"
 
-Posicion::Posicion(uint16_t x, uint16_t y, uint16_t anguloDeg) 
+Posicion::Posicion(float x, float y, uint16_t anguloDeg) 
     : x_(x)
     , y_(y)
     , anguloDeg_(anguloDeg) {
